@@ -19,7 +19,7 @@
 
 #define MIKTEX_COMP_MAJOR_VERSION 4
 #define MIKTEX_COMP_MINOR_VERSION 3
-#define MIKTEX_COMP_PATCH_VERSION 0
+#define MIKTEX_COMP_PATCH_VERSION 1
 
 #define MIKTEX_COMP_ORIG_VERSION_STR "2.11.1"
 
