@@ -19,7 +19,7 @@
 
 #define MIKTEX_COMP_MAJOR_VERSION 4
 #define MIKTEX_COMP_MINOR_VERSION 1
-#define MIKTEX_COMP_PATCH_VERSION 1
+#define MIKTEX_COMP_PATCH_VERSION 2
 
 #define MIKTEX_COMP_COPYRIGHT_STR "© 1996-2021 Christian Schenk"
 
